cfunc.h, clock_gettime.h, ftw.h, getopt.h, glob.h, pthread.cpp, sem.cpp, unistd.cpp

+++ unistd/sys/sys_types.h
@@ -14,7 +14,7 @@ typedef int sigval_t;
 typedef int sigset_t;
 typedef unsigned short ushort;
 typedef	int key_t;
-typedef int ssize_t;
+typedef intptr_t ssize_t;
 typedef unsigned short mode_t;
 typedef int gid_t;
 typedef int uid_t;
